bug fixes and refactors
